Fix exception due to crypto vs dist_tls race condition
Previously if the crypto module was loaded slowly in an Erlang node
using SSL, and the loading didn't finish when the dist_tls application
started, we got an exception.

IMPORTANT NOTE: This commit is known to break other aspects of
Erlang/OTP that we didn't need; use at your own risk.